feat: added vue-components/ NbButtonMechanical2
